Auth refactor
