+++ app/api/platform/ground-stations/ground-station-service.ts
@@ -1,4 +1,9 @@
+"use server";
+
 import { mockGroundStations, GroundStation } from "./mock";
+import { apiClient } from "@/lib/api-client";
+
+const API_PATH = '/ground-stations';
 
 export type GroundStationWithApiKey = GroundStation & {
   applicationId: string;
@@ -9,33 +14,14 @@ export type CreateGroundStationPayload = Omit<GroundStation, 'id' | 'createdAt' 
   isActive?: boolean;
 };
 
-const API_URL = 'http://localhost:5111/api/v1/ground-stations';
+export type UpdateGroundStationPayload = Partial<Omit<GroundStation, 'id' | 'createdAt' | 'isActive'>>;
 
 export async function getGroundStations(): Promise<GroundStation[]> {
-
   if (process.env.MOCKED || process.env.NEXT_PUBLIC_MOCKED) {
     return mockGroundStations;
   }
-
   try {
-    const response = await fetch(API_URL, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      next: { revalidate: 60 } 
-    });
-    if (response.status === 404) {
-      console.log("No flight plans found on the server, returning empty list.");
-      return [];
-    }
-    
-    if (!response.ok) {
-      throw new Error(`Failed to fetch flight plans: ${response.statusText}`);
-    }
-    
-    const data = await response.json();
-    return data || [];
+    return await apiClient.get<GroundStation[]>(API_PATH);
   } catch (error) {
     console.error("Error fetching ground stations:", error);
     return [];
@@ -59,50 +45,30 @@ export async function createGroundStation(
   }
 
   try {
-    const response = await fetch(API_URL, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(station),
-    });
-    
-    if (!response.ok) {
-        const errorBody = await response.text();
-        console.error('Failed to create ground station:', response.statusText, errorBody);
-        throw new Error(`Failed to create ground station: ${response.statusText}`);
-    }
-    return await response.json();
+    return await apiClient.post<CreateGroundStationPayload, GroundStationWithApiKey>(API_PATH, station);
   } catch (error) {
     console.error('Error creating ground station:', error);
     throw error;
   }
 }
 
-export async function updateGroundStation(updated: GroundStation): Promise<GroundStation> {
+export async function updateGroundStation(
+  id: number, 
+  payload: UpdateGroundStationPayload
+): Promise<GroundStation> {
   if (process.env.MOCKED || process.env.NEXT_PUBLIC_MOCKED) {
-  const idx = mockGroundStations.findIndex((g) => g.id === updated.id);
-  if (idx !== -1) {
-    mockGroundStations[idx] = { ...updated };
+    const idx = mockGroundStations.findIndex((g) => g.id === id);
+    if (idx !== -1) {
+      mockGroundStations[idx] = { ...mockGroundStations[idx], ...payload };
+    }
+    return mockGroundStations[idx];
   }
-  return updated;
-}
+
   try {
-    const response = await fetch(`${API_URL}/${updated.id}`, {
-      //TODO: make this a real patch :( its not right now
-      method: 'PATCH',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(updated),
-      next: { revalidate: 60 } 
-    });
-    
-    if (!response.ok) throw new Error(`Failed to update ground station: ${response.statusText}`);
-    return await response.json();
+    return await apiClient.patch<UpdateGroundStationPayload, GroundStation>(`${API_PATH}/${id}`, payload);
   } catch (error) {
-    console.error(`Error updating ground station ${updated.id}:`, error);
-    return updated; // Return original on error
+    console.error(`Error updating ground station ${id}:`, error);
+    throw error;
   }
 }
 
@@ -113,22 +79,16 @@ export async function deleteGroundStation(id: number): Promise<{ success: boolea
       mockGroundStations.splice(idx, 1);
       return { success: true };
     }
+    return { success: false };
   }
+
   try {
-    const response = await fetch(`${API_URL}/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      next: { revalidate: 60 } 
-    });
-    
-    if (!response.ok) throw new Error(`Failed to delete ground station: ${response.statusText}`);
+    await apiClient.delete(`${API_PATH}/${id}`);
     return { success: true };
   } catch (error) {
     console.error(`Error deleting ground station ${id}:`, error);
+    return { success: false };
   }
-  return { success: false };
 }
 
 // health check endpoint
@@ -159,20 +119,7 @@ export async function checkGroundStationHealth(id: number): Promise<GroundStatio
   }
   
   try {
-    const response = await fetch(`${API_URL}/${id}/health`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      cache: 'no-store' // Don't cache health checks as they should be real-time
-    });
-    
-    if (!response.ok) {
-      throw new Error(`Failed to check ground station health: ${response.statusText}`);
-    }
-    
-    const healthData = await response.json();
-    return healthData;
+    return await apiClient.get<GroundStationHealthResponse>(`${API_PATH}/${id}/health`);
   } catch (error) {
     console.error(`Error checking health of ground station ${id}:`, error);
     return null;
